+++ src/auth.js
@@ -29,10 +29,10 @@ const auth = () => {
   })
 }
 
-// code 换取 session_key
+// code 换取 session_key，生成并获取 3rd_session 即 token
 const sessionInit = (code, resolve, reject) => {
   return request({
-    url: API.INIT,
+    url: API.LOGIN,
     method: 'POST',
     data: {
       code: code
@@ -54,12 +54,6 @@ const sessionInit = (code, resolve, reject) => {
 
 const login = (userInfo = true) => {
   if (userInfo) {
-    if (storage.get(constants.STORAGE_KEY.USERINFO)) {
-      return new Promise(resolve => {
-        resolve(makeLoginResponseData())
-      })
-    }
-
     if (isLogining) {
       return new Promise((resolve, reject) => {
         loginResolve.push(resolve)
@@ -88,12 +82,6 @@ const login = (userInfo = true) => {
 }
 
 const silentLogin = () => {
-  if (storage.get(constants.STORAGE_KEY.UID)) {
-    return new Promise(resolve => {
-      resolve(makeLoginResponseData(false))
-    })
-  }
-
   if (isSilentLogining) {
     return new Promise((resolve, reject) => {
       silentLoginResolve.push(resolve)
@@ -162,9 +150,8 @@ const handleLoginFailure = (userInfo = true) => {
 }
 
 const logout = () => {
-  BaaS.check()
-
   return new Promise((resolve, reject) => {
+    // API.LOGOUT 接口不做 token 检查
     request({ url: API.LOGOUT, method: 'POST' }).then(() => {
       BaaS.clearSession()
       resolve()
@@ -188,7 +175,7 @@ const getUserInfo = () => {
         userInfo.id = storage.get(constants.STORAGE_KEY.UID)
         userInfo.openid = storage.get(constants.STORAGE_KEY.OPENID)
         userInfo.unionid = storage.get(constants.STORAGE_KEY.UNIONID)
-        return baasLogin(payload, resolve, reject, userInfo)
+        return getSensitiveData(payload, resolve, reject, userInfo)
       },
       fail: () => {
         reject(makeLoginResponseData(false))
@@ -197,10 +184,10 @@ const getUserInfo = () => {
   })
 }
 
-// 登录 BaaS
-const baasLogin = (data, resolve, reject, userInfo) => {
+// 上传 signature 和 encryptedData 等信息，用于校验数据的完整性及解密数据，获取 unionid 等敏感数据
+const getSensitiveData = (data, resolve, reject, userInfo) => {
   return request({
-    url: API.LOGIN,
+    url: API.AUTHENTICATE,
     method: 'POST',
     data: data
   }).then((res) => {