+++ src/baasRequest.js
@@ -11,9 +11,13 @@ const utils = require('./utils')
 // BaaS 网络请求，此方法能保证在已登录 BaaS 后再发起请求
 // eslint-disable-next-line no-unused-vars
 const baasRequest = function ({ url, method = 'GET', data = {}, header = {}, dataType = 'json' }) {
-  return auth.login(false).then(() => {
+  if (BaaS.getAuthToken()) {
     return request.apply(null, arguments)
-  })
+  } else {
+    return auth.silentLogin().then(() => {
+      return request.apply(null, arguments)
+    })
+  }
 }